moved animation importing to operator so that individual animations can be added
